# docs/swagger.yaml
basePath: /
definitions:
...
  structs.AddUser:
    properties:
      age:
        example: 9
        format: int64
        type: integer
      email:
        example: testing@gmail.com
        type: string
      password:
        example: "123456"
        type: string
      username:
        example: testing
        type: string
    type: object
  structs.AddUserResult:
    properties:
      age:
        example: 9
        format: int64
        type: integer
      email:
        example: testing@gmail.com
        type: string
      id:
        example: 1
        format: int64
        type: integer
      username:
        example: testing
        type: string
    type: object
...
  structs.DeleteUserResult:
    properties:
      message:
        example: Your account has been successfully deleted
        type: string
    type: object
...
  structs.LoginUser:
    properties:
      email:
        example: testing@gmail.com
        type: string
      password:
        example: "123456"
        type: string
    type: object
  structs.LoginUserResult:
    properties:
      token:
        example: result of generated token
        type: string
    type: object
...
  structs.UpdateUser:
    properties:
      email:
        example: testing1@gmail.com
        type: string
      username:
        example: testing1
        type: string
    type: object
  structs.UpdateUserResult:
    properties:
      age:
        example: 9
        format: int64
        type: integer
      email:
        example: testing@gmail.com
        type: string
      id:
        example: 1
        format: int64
        type: integer
      updated_at:
        example: date
        type: string
      username:
        example: testing
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a API to reach out my final assignment
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MyGram api
  version: "1.0"
paths:
...
  /users/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete by User ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.DeleteUserResult'
      security:
      - ApiKeyAuth: []
      summary: Delete an User
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update by json users
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.UpdateUserResult'
      security:
      - ApiKeyAuth: []
      summary: Update an users
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login by json users
      parameters:
      - description: Login user
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/structs.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.LoginUserResult'
      summary: Login an users
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: add by json users
      parameters:
      - description: Add user
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/structs.AddUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/structs.AddUserResult'
      summary: Add an users
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"